Turn all the CGFloats into floats Turn all the NSIntegers into integers

+++ gettysburgRecreation/XYZRegiment.h
@@ -11,13 +11,13 @@
 @interface XYZRegiment : NSObject
 
 // A float that keeps track of a regiment's health throughout
-@property CGFloat health;
+@property float health;
 
 // A label designating a regiment's allegiance
 @property NSString *allegiance;
 
 // The attacking power of a regiment
-@property CGFloat combatStrength;
+@property float combatStrength;
 
 // An array that holds the x and y location of the regiment, where
 // the integer portion is the specific tile and the decimal portion is
@@ -25,9 +25,9 @@
 @property NSMutableArray *locationArray;
 
 // Velocity in x and y direction of the regiment
-@property CGFloat xVelocity;
-@property CGFloat yVelocity;
-@property CGFloat maxSpeed;
+@property float xVelocity;
+@property float yVelocity;
+@property float maxSpeed;
 @property BOOL isAlive;
 
 // A method that moves the reigment
